USF_Commencement_Agent/USF_Commencement_Agent/requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
fpdf2>=2.8.0
